backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pymysql==1.1.0
redis==5.0.1
prometheus-client==0.19.0
psutil==5.9.6
cryptography==41.0.5
python-multipart==0.0.6 